// src/app/models/info-item-curso.ts
export class InfoItemCurso {

  constructor() {
    this.titulo = this.categoria = this.urlCurso = this.urlImagem =
      this.altImagem = this.descricaoItem = '';
  }

  titulo: string;
  categoria: string;
  urlCurso: string;
  descricaoItem: string;
  urlImagem: string;
  altImagem: string;
}